backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-core
langchain-openai
httpx
pydantic
langfuse
